# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisRead": {
        "properties": {
          "actors": {
            "items": {
              "$ref": "#/components/schemas/DiamondEntity"
            },
            "title": "Actors",
            "type": "array"
          },
          "ai_claims": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Ai Claims",
            "type": "array"
          },
          "ai_classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Classification"
          },
          "ai_relevance_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Relevance Score"
          },
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "ai_verification": {
            "additionalProperties": true,
            "title": "Ai Verification",
            "type": "object"
          },
          "analyst_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Note"
          },
          "attack_techniques": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Attack Techniques",
            "type": "array"
          },
          "automation_status": {
            "title": "Automation Status",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/DiamondEntity"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "confidence_auto": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Auto"
          },
          "content_text": {
            "title": "Content Text",
            "type": "string"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_coverage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Coverage"
          },
          "extraction_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Error"
          },
          "extraction_status": {
            "title": "Extraction Status",
            "type": "string"
          },
          "fetched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "final_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Url"
          },
          "infrastructure": {
            "items": {
              "$ref": "#/components/schemas/DiamondEntity"
            },
            "title": "Infrastructure",
            "type": "array"
          },
          "method_version": {
            "title": "Method Version",
            "type": "string"
          },
          "model_config_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Config Id"
          },
          "observables": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Observables",
            "type": "array"
          },
          "review_status": {
            "title": "Review Status",
            "type": "string"
          },
          "reviewed_actors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DiamondEntity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Actors"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_capabilities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DiamondEntity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Capabilities"
          },
          "reviewed_infrastructure": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DiamondEntity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Infrastructure"
          },
          "reviewed_victims": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DiamondEntity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Victims"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "victims": {
            "items": {
              "$ref": "#/components/schemas/DiamondEntity"
            },
            "title": "Victims",
            "type": "array"
          }
        },
        "required": [
          "extraction_status",
          "review_status",
          "content_text",
          "final_url",
          "content_type",
          "fetched_at",
          "extraction_error",
          "actors",
          "capabilities",
          "infrastructure",
          "victims",
          "evidence",
          "observables",
          "attack_techniques",
          "reviewed_actors",
          "reviewed_capabilities",
          "reviewed_infrastructure",
          "reviewed_victims",
          "confidence_auto",
          "method_version",
          "automation_status",
          "ai_relevance_score",
          "ai_classification",
          "ai_summary",
          "ai_claims",
          "ai_verification",
          "evidence_coverage",
          "decision_reason",
          "model_config_id",
          "analyst_note",
          "reviewed_at",
          "reviewed_by",
          "version",
          "updated_at"
        ],
        "title": "AnalysisRead",
        "type": "object"
      },
      "AnalysisRevisionRead": {
        "properties": {
          "analyst_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Note"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "report_id": {
            "format": "uuid",
            "title": "Report Id",
            "type": "string"
          },
          "review_version": {
            "title": "Review Version",
            "type": "integer"
          },
          "reviewed_by": {
            "title": "Reviewed By",
            "type": "string"
          },
          "snapshot": {
            "additionalProperties": true,
            "title": "Snapshot",
            "type": "object"
          }
        },
        "required": [
          "id",
          "report_id",
          "review_version",
          "decision",
          "snapshot",
          "analyst_note",
          "reviewed_by",
          "created_at"
        ],
        "title": "AnalysisRevisionRead",
        "type": "object"
      },
...
      "DiamondEntity": {
        "properties": {
          "confidence": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "integer"
          },
          "evidence": {
            "default": "",
            "maxLength": 5000,
            "title": "Evidence",
            "type": "string"
          },
          "name": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "type": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "confidence"
        ],
        "title": "DiamondEntity",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportDetail": {
        "properties": {
          "ai_classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Classification"
          },
          "ai_relevance_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Relevance Score"
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "confidence_auto": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Auto"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "extraction_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Status"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "relevance_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Relevance Reasons",
            "type": "array"
          },
          "relevance_score": {
            "title": "Relevance Score",
            "type": "integer"
          },
          "review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "source_name",
          "title",
          "canonical_url",
          "language",
          "summary",
          "relevance_score",
          "relevance_reasons",
          "status",
          "published_at",
          "created_at",
          "extraction_status",
          "review_status",
          "confidence_auto",
          "ai_classification",
          "ai_relevance_score",
          "analysis"
        ],
        "title": "ReportDetail",
        "type": "object"
      },
      "ReportSummary": {
        "properties": {
          "ai_classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Classification"
          },
          "ai_relevance_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Relevance Score"
          },
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "confidence_auto": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Auto"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "extraction_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Status"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "relevance_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Relevance Reasons",
            "type": "array"
          },
          "relevance_score": {
            "title": "Relevance Score",
            "type": "integer"
          },
          "review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "source_name",
          "title",
          "canonical_url",
          "language",
          "summary",
          "relevance_score",
          "relevance_reasons",
          "status",
          "published_at",
          "created_at",
          "extraction_status",
          "review_status",
          "confidence_auto",
          "ai_classification",
          "ai_relevance_score"
        ],
        "title": "ReportSummary",
        "type": "object"
      },
...
      "ReviewDecision": {
        "properties": {
          "actors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DiamondEntity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actors"
          },
          "analyst_note": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Note"
          },
          "capabilities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DiamondEntity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capabilities"
          },
          "confidence_analyst": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Analyst"
          },
          "decision": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "event_title": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Title"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "infrastructure": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DiamondEntity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infrastructure"
          },
          "reviewed_by": {
            "default": "local-analyst",
            "maxLength": 100,
            "minLength": 1,
            "title": "Reviewed By",
            "type": "string"
          },
          "victims": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DiamondEntity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Victims"
          }
        },
        "required": [
          "decision",
          "expected_version"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "APT Hunter API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "operationId": "list_review_queue_api_v1_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "review_status",
            "required": false,
            "schema": {
              "default": "pending",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ],
              "title": "Review Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportSummary"
                  },
                  "title": "Response List Review Queue Api V1 Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Review Queue",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{report_id}/decision": {
      "post": {
        "operationId": "decide_review_api_v1_reviews__report_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{report_id}/revisions": {
      "get": {
        "operationId": "list_review_revisions_api_v1_reviews__report_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRevisionRead"
                  },
                  "title": "Response List Review Revisions Api V1 Reviews  Report Id  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Review Revisions",
        "tags": [
          "reviews"
        ]
      }
    },
...
  }
}